feat(memory): Review B fixes for forgetting extension
- Remove unreachable ExecuteReturnedResults arm in init_schema migration
- Delete with_forgetting method and its tests (no production callers)
- Add doc comment to write_entries_to_db noting default TTL usage
- Change ForgettingConfig fields from Option<i32> to Option<i64>
- Remove as i64 casts throughout miner and active-searcher call chains

Source: Review B 问题修复
Type: refactor

+++ crates/config/src/agents/config_types/mod.rs
@@ -515,11 +515,11 @@ pub fn default_capacity_max_rules() -> usize {
 pub struct ForgettingConfig {
     /// Initial TTL in days for new events. `None` means inherit global default.
     #[serde(default)]
-    pub initial_ttl_days: Option<i32>,
+    pub initial_ttl_days: Option<i64>,
     /// Days to extend `expires_at` on active-searcher injection hit.
     /// `None` means inherit global default.
     #[serde(default)]
-    pub injection_extension_days: Option<i32>,
+    pub injection_extension_days: Option<i64>,
 }
 
 impl Default for ForgettingConfig {
@@ -544,12 +544,12 @@ impl ForgettingConfig {
 }
 
 /// Default initial TTL in days for new events.
-pub fn default_forgetting_initial_ttl_days() -> i32 {
+pub fn default_forgetting_initial_ttl_days() -> i64 {
     90
 }
 
 /// Default injection extension days for active-searcher hits.
-pub fn default_forgetting_injection_extension_days() -> i32 {
+pub fn default_forgetting_injection_extension_days() -> i64 {
     7
 }
